+++ src/components/Header/Header.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import './Header.css';
+
+export default function Header() {
+  return (
+    <div>
+      Math Problems Generator
+    </div>
+  )
+}
