+++ src/components/NumberOfProblems/NumberOfProblems.jsx
@@ -1,0 +1,19 @@
+import React, { useState } from 'react'
+
+export default function NumberOfProblems({ numberOfProblemsState }) {
+  const [numberOfProblems, setNumberOfProblems] = useState(10);
+
+  const updateNumberOfProblems = ({ target }) => {
+    setNumberOfProblems(Number(target.value));
+  };
+
+  return (
+    <div>
+      <input 
+        type="number" 
+        onChange={updateNumberOfProblems}
+        value={numberOfProblems}
+      />
+    </div>
+  )
+}
